Added the main file
This is the main file required for the programming of the ATmega328P. After power is applied the LED will light up. By pushing button 3 you can change which temperature sensor is used and this is indicated by an arrow on the screen. Button 2 increases the temperature by a single ˚C from a range of 10-35˚C. If the user attempts to go beyond 35˚C then the temperature loops back to 10˚C. Button 1 has the same functionality as button 2 but it decreases the temperature. If the selected temperature is within 1˚C deadband of the user temperature, then the device will continue the current mode. If the user temperature is 1˚C deadband higher than the current temperature of the sensor then the furnace will be activated, and the mode will switch from off to on. This is indicated by an LED.